refacto(select-location): Move form in a form component
