+++ pix-editor/app/components/form/select-location.hbs
@@ -1,0 +1,77 @@
+<form action="" class="ui form">
+  <PixSelect
+    @id="select-framework-location"
+    @value={{this.selectedFramework.value}}
+    @options={{this.frameworkList}}
+    @onChange={{this.selectFramework}}
+    @hideDefaultOption={{true}}
+  >
+    <:label>Référentiel</:label>
+  </PixSelect>
+  <PixSelect
+    @id="select-competence-location"
+    @value={{this.selectedCompetence.value}}
+    @options={{this.competenceList}}
+    @onChange={{this.selectCompetence}}
+    @hideDefaultOption={{true}}
+  >
+    <:label>Compétence</:label>
+  </PixSelect>
+  {{#if @isMovingTube}}
+    <PixSelect
+      data-test-select-thematic
+      @id="select-thematic-location"
+      @value={{this.selectedTheme}}
+      @options={{this.themeList}}
+      @onChange={{this.selectTheme}}
+      @hideDefaultOption={{true}}
+      @placeholder="Sélectionner une thématique"
+    >
+      <:label>Thématique*</:label>
+    </PixSelect>
+  {{/if}}
+  {{#if @selectTubeLevel}}
+    <PixSelect
+      data-test-select-tube
+      @id="select-tube-location"
+      @value={{this.selectedTube}}
+      @options={{this.tubeList}}
+      @onChange={{this.selectTube}}
+      @hideDefaultOption={{true}}
+      @placeholder="Sélectionner un sujet"
+    >
+      <:label>Sujet</:label>
+    </PixSelect>
+    {{#if this.selectedTube}}
+      {{#if @isMovingPrototype}}
+        <div data-test-skill-list class="field">
+          {{#if (gt this.skillsListWithCategory.length 0) }}
+            <PixSelect
+              data-test-select-skill
+              @id="select-skill-location"
+              @value={{this.selectedSkill}}
+              @options={{this.skillsListWithCategory}}
+              @onChange={{this.selectSkill}}
+              @hideDefaultOption={{true}}
+              @placeholder="Sélectionner un acquis"
+            >
+              <:label>Acquis</:label>
+            </PixSelect>
+          {{/if}}
+        </div>
+      {{else}}
+        <PixSelect
+          data-test-select-level
+          @id="select-level-location"
+          @value={{this.selectedLevel}}
+          @options={{this.selectLevelOptions}}
+          @onChange={{this.selectLevel}}
+          @hideDefaultOption={{true}}
+          @placeholder="Sélectionner un niveau"
+        >
+          <:label>Niveau</:label>
+        </PixSelect>
+      {{/if}}
+    {{/if}}
+  {{/if}}
+</form>
