+++ pix-editor/app/components/pop-in/select-location.hbs
@@ -1,41 +1,23 @@
-<PixModal @title={{this.titleModal}}
-          @onCloseButtonClick={{this.closeModal}}
-          @showModal={{@showModal}}
+<PixModal
+  @title={{this.titleModal}}
+  @onCloseButtonClick={{this.closeModal}}
+  @showModal={{@showModal}}
 >
   <:content>
-    <form action="" class="ui form">
-      <Field::Select data-test-select-source @title="Référentiel" @value={{this.selectedFramework}} @options={{this.frameworkList}} @edition={{true}} @setValue={{this.selectFramework}} />
-      <Field::Select data-test-select-competence @title="Compétence" @value={{this.selectedCompetence}} @options={{this.competenceList}} @edition={{true}} @setValue={{this.selectCompetence}}/>
-      {{#if @isTubeLocation}}
-        <Field::Select data-test-select-theme @title="Thématique*" @defaultText="Sélectionner une thématique" @value={{this.selectedTheme}} @options={{this.themeList}} @edition={{true}} @setValue={{this.selectTheme}}/>
-      {{/if}}
-      {{#if @selectTubeLevel}}
-        <Field::Select @title="Tube" @value={{this.selectedTube}} @options={{this.tubeList}} @edition={{true}} @setValue={{this.selectTube}}/>
-        {{#if this.selectedTube}}
-          {{#if @isPrototypeLocation}}
-            <div data-test-skill-list class="field">
-              <label>Acquis</label>
-              <PowerSelect  @options={{this.skillsGroupByLevelList}} @searchEnabled={{false}} @selected={{this.selectedSKill}} @onChange={{this.selectSkill}} as |skill|>
-                <span class="skill-list">{{skill.name}} (v.{{skill.version}}) <span class="skill-list__status {{if (eq skill.status "actif") "active" "draft"}}" title={{skill.status}}></span></span>
-              </PowerSelect>
-            </div>
-          {{else}}
-            <Field::Select @title="Niveau" data-test-select-level @value={{this.selectedLevel}} @options={{this.selectLevelOptions}} @edition={{true}} @setValue={{this.selectLevel}}/>
-          {{/if}}
-        {{/if}}
-      {{/if}}
-    </form>
+    <Form::SelectLocation />
   </:content>
   <:footer>
-    <PixButton @backgroundColor="transparent-light"
-               @isBorderVisible={{true}}
-               @triggerAction={{this.closeModal}}
+    <PixButton
+      @backgroundColor="transparent-light"
+      @isBorderVisible={{true}}
+      @triggerAction={{this.closeModal}}
     >
       {{t 'common.cancel'}}
     </PixButton>
-    <PixButton data-test-move-action
-               @triggerAction={{this.setLocation}}
-               @isDisabled={{not this.enableMoveActionButton}}
+    <PixButton
+      data-test-move-action
+      @triggerAction={{this.setLocation}}
+      @isDisabled={{not this.enableMoveActionButton}}
     >
       <i class="checkmark icon"></i>
       {{this.titleButtonAction}}
