+++ pix-editor/app/templates/authenticated/competence/prototypes/single.hbs
@@ -160,10 +160,9 @@
     @onChange={{this.setSkill}}
     @title="Acquis du prototype"
     @selectTubeLevel={{true}}
-    @isPrototypeLocation={{true}}
+    @isMovingPrototype={{true}}
     @tube={{this.challenge.skill.tube}}
     @skill={{this.challenge.skill}}
-    @multipleLevel={{true}}
     @close={{this.closeMovePrototype}}
     @showModal={{this.displaySelectLocation}}
   />
