+++ pix-editor/app/templates/authenticated/competence/skills/single.hbs
@@ -104,9 +104,7 @@
         @title={{concat "Destination de la copie de " this.skillName}}
         @selectTubeLevel={{true}}
         @tube={{this.skill.tube}}
-        @level={{this.skill.level}}
-        @selectEmptyLevels={{true}}
-        @isSkillLocation={{true}}
+        @isMovingSkill={{true}}
         @close={{this.closeSelectLocation}}
         @showModal={{this.displaySelectLocation}}
 />
