+++ pix-editor/app/templates/authenticated/competence/tubes/single.hbs
@@ -57,7 +57,7 @@
   @onChange={{this.setCompetence}}
   @name={{this.tube.name}}
   @theme={{this.tube.theme}}
-  @isTubeLocation={{true}}
+  @isMovingTube={{true}}
   @close={{this.closeSelectCompetence}}
   @showModal={{this.displaySelectLocation}}
 />
